Добавлены меню и музыка

+++ SFML5/globalVariableWithGame.h
@@ -3,7 +3,7 @@
 
 struct globalVariable {
 	int g_pointRessurection;
-	int g_timer;
+	float g_timer;
 	float g_currentFrame; // todo Создать вектор для CurrentFrame и запихуть их туда все!
 	float g_timerAppearenceEnemy;
 	float g_curFrame;
@@ -12,7 +12,7 @@ struct globalVariable {
 	int g_numberBonus;
 	globalVariable() {
 		g_pointRessurection = 0;
-		g_timer = 0;
+		g_timer = 0.0f;
 		g_currentFrame = 0.0f; // todo Создать вектор для CurrentFrame и запихуть их туда все!
 		g_timerAppearenceEnemy = 0.0f;
 		g_curFrame = 0.0f;
